LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/uievents/mouse/mouseevent_move_button-expected.txt

Mouse Events

Test Description: This test checks if mouse events set button property correctly
Put your mouse over the green rectangle
Press a non-primary button and hold it
Drag mouse to blue rectangle
Release mouse button


FAIL mouse events fired without button state assert_equals: Button must be un-initialized for mouseleave. expected 0 but got 1
